server/requirements.txt
torch==1.0.0
scikit-image==0.16.2
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
autopep8==1.5.1
bottle==0.12.18
certifi==2020.4.5.1
chardet==3.0.4
CProfileV==1.0.7
cycler==0.10.0
decorator==4.4.2
entrypoints==0.3
face-alignment==1.0.0
flake8==3.7.9
idna==2.9
imageio==2.8.0
kiwisolver==1.2.0
mccabe==0.6.1
networkx==2.4
numpy==1.18.2
Pillow==7.1.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
pyprof2calltree==1.4.4
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
snakeviz==2.0.1
tornado==6.0.4
tqdm==4.45.0
urllib3==1.25.8
